requirements.txt
asgiref==3.4.1
Django==3.2.9
pytz==2021.3
python-decouple==3.4
sqlparse
whitenoise
django-cors-headers
django-heroku
django-rest-framework
djangorestframework
gunicorn==20.1.0
dj-database-url
db-sqlite3